ibutils: drop unused parameter in OSM_LOG_ENTER macro
This follows named OpenSM patch (in master only):

__func__ macro is used in the OSM_LOG_ENTER() to show an actual function
name, so the second parameter is not really useful here. OTOH it makes
it harder to grep over OpenSM source code, when searches are by function
names it generates a lot of unrelated matches.

Signed-off-by: Sasha Khapyorsky <sashak@voltaire.com>

orenk: NOTE: with this commit, ibis seg faults when binding to port when tested on my machine -
       This commit is pushed for compilation - still to be debugged.