ibutils: ibdiagnet -r "Dead end" errors
On a cluster running sles11 and OFED 1.4, we recently started seeing errors
like this:

# ibdiagnet -r
.....
-I-
-I- Verifying all CA to CA paths ...
-E- Unassigned LFT for lid:1 Dead end at:S0800690000004057/U1
-E- Fail to find a path from:r1i0n9/U1/1 to:r1lead/U1/1
...

But the forwarding tables (obtained with dump_lfts.sh, and smpdump)
are correct. The problem turned out to be that the string "-lft" was
being interpreted as a port number, resulting in an off-by-one error.

The following fixed it for us.

Signed-off-by: Arthur Kepner <akepner@sgi.com>
Signed-off-by: Yevgeny Kliteynik <kliteyn@dev.mellanox.co.il>