Added an option to use the from edge new topo match algorithm.
The from edge algorithm and the old algorithm that used
to be excuted are mutually exclusive, this means that
only one of them can be use to do topology match.

The from edge algorithm isn't exposed to all users yet
(in the help information). It can be used with the -e option.
The new algorithm can be used when running from management
switch (port0) instead of the old algorithm.

Signed-off-by: Aviad Yehezkel <aviadye@mellanox.co.il>
Signed-off-by: Yevgeny Kliteynik <kliteyn@dev.mellanox.co.il>