Congestion.cpp: bug fix
When printing out the port HSD we can not use the
congData.portNumPaths as it is set by CongZero which
clears the list of paths...
So instead we simply out the length of the path list.
This reverts an early modification intended to output the
calculated HSD that takes into account the impact of the first
link that the paths merge on and then each one of them gets
just a fraction of the BW (so it avoid double counting
of number of HSD links).

Signed-off-by: Eitan Zahavi <eitan@mellanox.co.il>
Signed-off-by: Yevgeny Kliteynik <kliteyn@dev.mellanox.co.il>